Update encoder logic before finalising creation of post

+++ backend/app/server/controllers/post.py
@@ -1,13 +1,15 @@
 import cv2
+import numpy as np
 import tempfile
+from PIL import Image
 from bson.objectid import ObjectId
 from fastapi import HTTPException
 from ..database import posts_collection, users_collection
 from .like import get_all_likes_on_post
 from .comment import get_all_comments_on_post
 from .steganography.hash import hash_sha_info
-from .upload import upload_image_path
-from .verify import encode_image
+from .upload import upload_image_path, upload_image
+from .verify import encode_image, decode_image
 
 # helpers
 
@@ -40,14 +42,28 @@ async def initialize_post(user_id: ObjectId, post: dict, image_url: str, user_sh
 # Add a new post into to the database
 async def add_post(email: str, post_data: dict) -> dict:
     user = await users_collection.find_one({"email": email})
-    user_sha, info = hash_sha_info(str(user["_id"]))
+    upload_file = post_data["image"]
 
-    encoded_image = await encode_image(post_data["image"], info)
+    image_url = upload_image(upload_file)
 
-    # upload the image
-    with tempfile.NamedTemporaryFile(suffix=".png") as tmp:
-        cv2.imwrite(tmp.name, encoded_image)
-        image_url = upload_image_path(tmp.name)
+    pil_image = Image.open(upload_file.file)
+    rgb_image = cv2.cvtColor(np.array(pil_image), cv2.COLOR_RGB2BGR)
+
+    embedded_sha = await decode_image(rgb_image)
+
+    sha_already_exists = await posts_collection.find_one({"user_sha": embedded_sha})
+
+    if not sha_already_exists:
+        user_sha, info = hash_sha_info(str(user["_id"]))
+        encoded_image = await encode_image(rgb_image, info)
+
+        # upload the image
+        with tempfile.NamedTemporaryFile(suffix=".png") as tmp:
+            cv2.imwrite(tmp.name, encoded_image)
+            image_url = upload_image_path(tmp.name)
+
+    else:
+        user_sha = sha_already_exists["user_sha"]
 
     # delete file
     del post_data["image"]
